Introduced MediaAssets Generator
